src/main/webapp/app/entities/cbdc-account/delete/cbdc-account-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, inject, fakeAsync, tick } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { CBDCAccountService } from '../service/cbdc-account.service';

import { CBDCAccountDeleteDialogComponent } from './cbdc-account-delete-dialog.component';

describe('CBDCAccount Management Delete Component', () => {
  let comp: CBDCAccountDeleteDialogComponent;
  let fixture: ComponentFixture<CBDCAccountDeleteDialogComponent>;
  let service: CBDCAccountService;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule],
      declarations: [CBDCAccountDeleteDialogComponent],
      providers: [NgbActiveModal],
    })
      .overrideTemplate(CBDCAccountDeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(CBDCAccountDeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(CBDCAccountService);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete(123);
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith(123);
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      })
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
